splited to components. CORS. Mime-types
