+++ package.json
@@ -1,15 +1,16 @@
 {
-  "name": "lec-1",
+  "name": "saber_project",
   "version": "1.0.0",
-  "description": "",
+  "description": "vs.ru alike web-spa",
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "keywords": [],
-  "author": "",
+  "author": "Denis, Svetlana, Yuriy, Alexey",
   "license": "ISC",
   "dependencies": {
-    "debug": "^4.3.2"
+    "debug": "^4.3.2",
+    "mime": "^2.5.2"
   }
 }
