+++ public/index.html
@@ -4,12 +4,13 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
 
-    <title>Sample</title>
+    <title>Sh</title>
     <link rel="stylesheet" href="./main.css" />
 </head>
 <body>
     <div id="root"></div>
 
-    <script src="./main.js"></script>
+    <script type="module" src="./main.js"></script>
+    <!-- <script src="./main.js"></script> -->
 </body>
 </html>
