+++ public/main.js
@@ -1,4 +1,6 @@
-'use strict';
+// 'use strict';
+
+import {switchIsRegistered, signupPage, authCfg} from "./components/authorizationForm.js";
 
 // TODO: разбить по функциональным элементам, выделить конфиг
 
@@ -25,7 +27,7 @@ const configuration = {
     open: [
       () => {
         console.log("set IsRegistered to false");
-        sessionParams.isRegistered = false;
+        authCfg.isRegistered = false;
       },
       signupPage
     ]
@@ -36,7 +38,7 @@ const configuration = {
     open: [
       () => {
         console.log("set IsRegistered to true");
-        sessionParams.isRegistered = true;
+        authCfg.isRegistered = true;
       },
       signupPage
     ]
@@ -48,17 +50,13 @@ const configuration = {
   },
 
   // others (no menu)
-  changeRegFormTypeBtn: {
+  changeRegFormType: {
     open: [
       switchIsRegistered,
       signupPage
     ]
   },
-}
-
-const sessionParams = {
-  isRegistered: false,
-}
+};
 
 /////////////////////////////////////
 //
@@ -66,30 +64,18 @@ const sessionParams = {
 //
 /////////////////////////////////////
 
-function switchIsRegistered() {
-  sessionParams.isRegistered = !sessionParams.isRegistered;
-  console.log("IsRegistered switched to " + sessionParams.isRegistered);
-}
-
-function createInput(type, placeholder, name) {
-  const input = document.createElement('input');
-  input.type = type;
-  input.name = name;
-  input.placeholder = placeholder;
-
-  return input;
-}
-
 ////////////////////////////////
-
+//
 //              Pages
-
+//
 ////////////////////////////////
 
 function mainPage() {
   // TODO: убрать в хедер и сайд-бар
 
   root.innerHTML = "";
+  // root.title = "SaberProject";
+  document.title = "SaberProject";
   menuElements.map( (menu_el) => {
     if (!(menu_el in configuration)) {
       console.log("Error: " + menu_el + "is not described in configuration");
@@ -108,64 +94,6 @@ function mainPage() {
   });
 }
 
-
-function signupPage() {
-  let isRegistered = sessionParams.isRegistered;
-  // стираем старые элементы, чтобы нарисовать новые
-  root.innerHTML = "";
-
-  const form = document.createElement('form');
-
-  // поля формы
-  let emailInput = null;
-  let passwordRepeatInput = null;
-  const loginInput = createInput("login", "Логин", "login");
-  const passwordInput = createInput("password", "Пароль", "password");
-  if (!isRegistered) {
-    emailInput = createInput("email", "e-mail", "email");
-    passwordRepeatInput = createInput("password", "Повторите пароль", "passwordRepeat");
-  }
-
-  if (!isRegistered) form.appendChild(emailInput);
-  form.appendChild(loginInput);
-  form.appendChild(passwordInput);
-  if (!isRegistered) form.appendChild(passwordRepeatInput);
-
-
-  // интерефейс формы
-  const submitBtn = document.createElement('input');
-  submitBtn.type = 'submit';
-  submitBtn.value = "Зарегистрироваться";
-
-  const backBtn = document.createElement('a');
-  backBtn.textContent = "Назад";
-  backBtn.href = '/';
-  backBtn.dataset.section = 'main';
-  // backBtn.addEventListener("click", (e) => {
-  //   e.preventDefault();
-  //   // перерисовываем меню заново
-  //   mainPage();
-  // });
-
-  // const changeFormTypeBtn = document.createElement('button');
-  // changeFormTypeBtn.type = "button";
-  const changeFormTypeBtn = document.createElement('button');
-  changeFormTypeBtn.textContent = isRegistered ? "Создать аккаунт" : "У меня уже есть аккаунт";
-  changeFormTypeBtn.href = isRegistered ? "/login" : "/register";
-  changeFormTypeBtn.dataset.section = "changeRegFormType"
-  // changeFormTypeBtn.addEventListener("click", (e) => {
-  //   sessionParams.isRegistered = !isRegistered;
-  //   signupPage()
-  // })
-
-  form.appendChild(submitBtn);
-  form.appendChild(backBtn);
-  form.appendChild(changeFormTypeBtn);
-
-  // форма
-  root.appendChild(form);
-}
-
 ////////////////////////////////
 //
 //       Сама страница
@@ -173,14 +101,6 @@ function signupPage() {
 ////////////////////////////////
 
 mainPage()
-
-// const signupLink = document.querySelector('[data-section="signup"]')
-// signupLink.addEventListener('click', (e) => {
-//   // ссылка больше не кликается (это было ее поведение по умолчанию)
-//   e.preventDefault();
-//   // рисуем нужную страницу по клику
-//   signupPage(false);
-// });
 
 ////////////////////////////////
 //
@@ -193,19 +113,9 @@ root.addEventListener('click', e => {
     if (target instanceof HTMLAnchorElement) { // проверям, что клик был по ссылке (anchor)
           e.preventDefault();
 
-          const actions_arr = configuration[target.dataset.section].open
-          // TODO: check for undef
-          // TODO: fix that
-          console.log("targeter: " + target.href + " | " + actions_arr.lenght);
+          console.log("targeter: ", target.dataset.section)
 
-          actions_arr[0]();
-          actions_arr[1]();
-          // если target.dataset.section-ссылка описана в configuration, и у нее есть метод open, то он будет вызван
-          for (let i = 0; i < actions_arr.lenght; i++) {
-            configuration[target.dataset.section]?.open[i]();
-            console.log(i + ": " + configuration[target.dataset.section].open[i])
-          }
-          
+          configuration[target.dataset.section]?.open.map(action => action())
       }
   })
