+++ server/index.js
@@ -1,10 +1,12 @@
 const http = require('http');  // для работы с http
 const fs = require('fs');  // для работы с файловой системой
+const mime = require('mime/lite');  
 const debug = require('debug');
 
 const log = debug('*')
 
 const page404 = fs.readFileSync('./public/404.html')  // мы знаем, что код тут не упадет, т.к. страница существует
+const CORS = '*'
 
 const server = http.createServer((req, res) => {
     // забираем урл загружаемого документа из GET запроса
@@ -18,7 +20,16 @@ const server = http.createServer((req, res) => {
         if (err) {
             // вот тут отдать страничку 404
             data = page404;
+            // res.writeHead(404);
+        } else {
+            // res.writeHead(200);
         }
+
+        console.log("requested: ", path, "| ext: ", path.slice(path.lastIndexOf('.') + 1), " mime: ",  mime.getType(path.slice(path.lastIndexOf('.') + 1)))
+
+        res.setHeader("Content-Type", mime.getType(path.slice(path.lastIndexOf('.') + 1)));
+        // CORS
+        res.setHeader("Access-Control-Allow-Origin", CORS);
 
         res.write(data);
         res.end();
